refactor build-lib to mjs
