+++ package.json
@@ -119,7 +119,7 @@
     "build:js:typeless": "npm-run-all build:lib build:companion build:svelte",
     "build:js": "npm-run-all build:js:typeless build:locale-pack build:angular build:bundle",
     "build:ts": "yarn tsc -b && yarn workspace @uppy/svelte check",
-    "build:lib": "yarn node ./bin/build-lib.js",
+    "build:lib": "yarn node ./bin/build-lib.mjs",
     "build:locale-pack": "yarn workspace @uppy-dev/locale-pack build && eslint packages/@uppy/locales/src/en_US.ts --fix && yarn workspace @uppy-dev/locale-pack test unused",
     "build": "npm-run-all --serial build:ts --parallel build:js build:css --serial size",
     "contributors:save": "yarn node ./bin/update-contributors.mjs",
